+++ py_wake/utils/area_overlapping_factor.py
@@ -1,5 +1,6 @@
 from numpy import newaxis as na
 
+import py_wake
 import numpy as np
 
 
@@ -71,7 +72,7 @@ class AreaOverlappingFactor():
             R2 = np.zeros_like(R1) + R2
         if d.shape != R1.shape:
             d = np.zeros_like(R1) + d
-        A_ol_f = np.zeros(np.maximum(R1.shape, R2.shape))
+        A_ol_f = np.zeros(np.maximum(R1.shape, R2.shape), dtype=py_wake.dtype)
         p = (R1 + R2 + d) / 2.0
 
         # make sure R_big >= R_small
